+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased][unreleased]
 
+- New API methods for Database:
+  - async count(table, condition): number
 - Add typings for Database and Query classes
 
 ## [1.1.5][] - 2021-07-04