+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add typings for Database and Query classes
 - Move types to package root
 - Generated `always as identity` just for `Registry`
+- Support `DEFAULT CURRENT_TIMESTAMP` for `datetime`
 
 ## [1.1.5][] - 2021-07-04