+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Support `DEFAULT CURRENT_TIMESTAMP` for `datetime`
 - Generate SQL for bootstrapping db data
 - Pass `Model` to `Database` to access schemas
+- Use common `Console` interface for logging
 
 ## [1.1.5][] - 2021-07-04