+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Generate SQL for bootstrapping db data
 - Pass `Model` to `Database` to access schemas
 - Use common `Console` interface for logging
+- Generate ids for `Registry` tables
 
 ## [1.1.5][] - 2021-07-04